refactor: extract wizard into dedicated files and fix variable scoping
